+++ source/redux/server/render.js
@@ -5,20 +5,31 @@ import { ReduxRouter } from 'redux-router'
 import react_render_on_server from '../../render on server'
 
 import { location_url } from '../../location'
+import timer from '../../timer'
+
+function timed_react_render_on_server(named_arguments)
+{
+	const render_timer = timer()
+	const markup = react_render_on_server(named_arguments)
+	const result =
+	{
+		content : markup,
+		time    : render_timer()
+	}
+	return result
+}
 
 // Returns a Promise resolving to { status, content, redirect }.
 //
-export default function render_on_server({ monitoring, disable_server_side_rendering, create_page_element, render_webpage_as_react_element, url, store })
+export default function render_on_server({ disable_server_side_rendering, create_page_element, render_webpage_as_react_element, url, store })
 {
-	// HTTP response status code
-	let http_status_code
-
-	const routing_finished = monitoring.started('request.route')
+	// Routing only takes a couple of milliseconds
+	// const routing_finished = monitoring.started('route')
 
 	// Perform routing for this `url`
 	return match_url(url, store).then(routing_result =>
 	{
-		routing_finished()
+		// routing_finished()
 
 		// Return in case of an HTTP redirect
 		if (routing_result.redirect)
@@ -27,9 +38,16 @@ export default function render_on_server({ monitoring, disable_server_side_rende
 		}
 
 		// Http response status code
-		http_status_code = get_http_response_status_code_for_the_route(routing_result.matched_routes)
+		const http_status_code = get_http_response_status_code_for_the_route(routing_result.matched_routes)
 
-		const preloading_finished = monitoring.started('request.preload')
+		// Concatenated `react-router` route string.
+		// E.g. "/user/:user_id/post/:post_id"
+		const route = routing_result.matched_routes.map(route => route.path.replace(/^\//, '').replace(/\/$/, '')).join('/')
+
+		// Profiling
+		const time = {}
+
+		const preload_timer = timer()
 
 		// When `url` matching process finished,
 		// it immediately launched preloading process,
@@ -39,27 +57,28 @@ export default function render_on_server({ monitoring, disable_server_side_rende
 		//
 		return wait_for_page_to_preload(store).then(() => 
 		{
-			preloading_finished()
+			time.preload = preload_timer()
 
 			if (disable_server_side_rendering)
 			{
 				// Render the empty <Html/> component into Html markup string
-				return { content: react_render_on_server({ render_webpage_as_react_element }) }
+				const rendered = timed_react_render_on_server({ render_webpage_as_react_element })
+				time.render = rendered.time
+
+				// return  HTML markup
+				return { content: rendered.content, route, time }
 			}
 
 			// Renders the current page React component to a React element
 			// (`<ReduxRouter/>` is gonna get the matched route from the `store`)
 			const page_element = create_page_element(<ReduxRouter/>, { store })
 
-			const rendering_finished = monitoring.started('request.render')
-
 			// Render the current page's React element to HTML markup
-			const content = react_render_on_server({ render_webpage_as_react_element, page_element })
-
-			rendering_finished()
+			const rendered = timed_react_render_on_server({ render_webpage_as_react_element, page_element })
+			time.render = rendered.time
 
 			// return HTTP status code and HTML markup
-			return { status: http_status_code, content }
+			return { status: http_status_code, content: rendered.content, route, time }
 		})
 	})
 	.catch((error) =>